Create MinimizeHeightsII.cpp
Follow the steps below to solve the given problem:



Sort the array 
Try to make each height of the tower maximum by decreasing the height of all the towers to the right by k and increasing all the height of the towers to the left by k. Check whether the current index tower has the maximum height or not by comparing it with a[n]-k. If the tower's height is greater than the a[n]-k then it's the tallest tower available.
Similarly, find the shortest tower and minimize the difference between these two towers.